feat:> 팀 이슈 디테일 조회 ing
